[cli] move ParseArgs to cli module
